requirements.txt
click==8.0.1
colorama==0.4.4
colormap==1.0.3
cycler==0.10.0
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.15.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
Pillow==8.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
six==1.16.0
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
